+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @reservoir0x/relay-kit-ui
 
+## 5.0.1
+
+### Patch Changes
+
+- e43a756: Improved copy for tx steps
+
+## 5.0.0
+
+### Major Changes
+
+- d38abdf: Revamped transaction steps
+
+### Patch Changes
+
+- 3bd044c: Fix duplicate deposit success event
+- Updated dependencies [d38abdf]
+  - @relayprotocol/relay-sdk@3.0.0
+  - @relayprotocol/relay-kit-hooks@2.0.6
+
+## 4.0.19
+
+### Patch Changes
+
+- f54d1f3: Save recent custom addresses to local storage
+- cda3f6d: Use fill time on success screen
+- Updated dependencies [5135cd1]
+  - @relayprotocol/relay-sdk@2.4.6
+  - @relayprotocol/relay-kit-hooks@2.0.5
+
 ## 4.0.18
 
 ### Patch Changes